# Rewrote flatten & deflatten