As part of working on #8