@tmt96 2018-05-02
Performance result on examples (avg three runs):

|   | Master(s) | Flatten(s) |
| ---- | --- | ----- |
| Primes (to 10mil) | 18.080 | 16.236|
| Mandelbrot | 13.600 | 12.197|